see summary IMPORTANT
BY having at least two category objects, one either temporary or denoting a new category being creating and one for a category we are editing, we can keep one null at all times and check for this significance. THIS allows me to pass the correct intent key along with the relavant intent extras and these extras can also then be checked as we move to a new activity in order to make the appropraite aforementioned pair of variables null or not. THE CREATE CATEGORY ACTIVITY class initially checks if we're editing a category, if not, are we passing a temporary category back from the CreateDataEntry activity and if not, initialise the temporary category so we know which intents are to be passed around. THE CREATE DATA ENTRY ACTIVITY class only needs to check whether we're editting a category or creating a new category, the relevant intent keys will be pased back and fourth as required.
-
THIS STILL NEEDS TO BE TESTED FOR EDIT BUTTONS AS THE LIST ELEMENT EDIT BUTTONS HAVE YET TO BE CREATED FOR BOTH ADAPTERS FOR CREATING CATEGORIES AND DATA ENTRIES BUT CREATING NEW OBJECTS CURRENTLY WORKS, IMPLEMENTATION EXISTS FOR EDITING BUTTONS TO WORK BUT BUTTONS NOT CODED AT THIS POINT